+++ CHANGELOG.md
@@ -4,7 +4,7 @@ This list is not intended to be all-encompassing - it will document major and br
 changes with their rationale when appropriate:
 
 ### v4.12.0.1
-- **http4k-*** : Fix AWS event formats with wrong cases.
+- **http4k-*** : Fix #652 - AWS event format adapters have fields with wrong cases.
 
 ### v4.12.0.0
 - **http4k-*** : Upgrade some dependency versions, including Kotlin to 1.5.30.